MOD db.js error #1
